+++ CHANGELOG.md
@@ -132,3 +132,16 @@
 ## 3.3.3
 
 * Fix build script of Android plugin and remove AGP version requirement (#110)
+
+## 3.4.0
+
+* Add support for reading / write MIFARE Classic / Ultralight tags on Android (merged #82, partially fixes #82)
+* Add support for reading / write ISO 15693 tags on iOS (merged #117, partially fixes #68)
+* Fix compiling issues (#123)
+* Other minor fixes (#114, #115)
+
+## 3.4.1
+
+* Fix & split examples to example/ dir
+* Publish examples to pub.dev
+* Support transceiving of raw ISO15693 commands on iOS